coworkers/vsm_project/requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
CacheControl==0.12.11
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cleo==2.0.1
crashtest==0.4.1
deepl==1.14.0
distlib==0.3.6
dnspython==2.3.0
dulwich==0.20.50
filelock==3.9.0
frozenlist==1.3.3
html5lib==1.1
idna==3.4
importlib-metadata==6.0.0
jaraco.classes==3.2.3
jsonschema==4.17.3
keyring==23.13.1
lockfile==0.12.2
magic-filter==1.0.9
mongoengine==0.27.0
more-itertools==9.1.0
msgpack==1.0.4
multidict==6.0.4
packaging==23.0
pexpect==4.8.0
pipenv==2023.2.18
pkginfo==1.9.6
platformdirs==2.6.2
poetry==1.3.2
poetry-core==1.4.0
poetry-plugin-export==1.3.0
prettytable==3.6.0
ptyprocess==0.7.0
pycparser==2.21
pymongo==4.3.3
pyrsistent==0.19.3
python-dotenv==1.0.0
pytz==2022.7.1
rapidfuzz==2.13.7
requests==2.28.2
requests-toolbelt==0.10.1
shellingham==1.5.0.post1
six==1.16.0
tomlkit==0.11.6
trove-classifiers==2023.2.20
urllib3==1.26.14
virtualenv==20.19.0
virtualenv-clone==0.5.7
wcwidth==0.2.6
webencodings==0.5.1
xattr==0.10.1
yarl==1.8.2
zipp==3.15.0
